Moved schematic changes into PCB
All the changes made in the schematic have been moved into the PCB editor. It was a bit of work due to the amount of reorganizing done in the schematic which led to a lot of new reference numbers. Was easy enough to get most things back into place where they were before.